colour scale constructor for USSC colours

colour scale constructor for USSC colours

scale_colour_ussc(palette = "main", discrete = TRUE, reverse = FALSE, ...)

Arguments

palette
Character name of palette in ussc_pal (i.e. main, blue, light, dark, grey, mixed)

discrete
Boolean indicating whether colour aesthetic is discrete or not

reverse
Boolean indicating whether the palette should be reversed

...
Additional arguments passed to discrete_scale() or scale_colour_gradientn(), used respectively when discrete is TRUE or FALSE

Examples

Colour by discrete variable using default palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4) + scale_colour_ussc()
Reverse colour using blue palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4) + scale_colour_ussc('blue', reverse=T)
Remember, you can change the transparency of the colour by adding alpha to the geom_...() call
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) + scale_colour_ussc('blue', reverse=T)

#> Error: <text>:1:8: unexpected symbol
#> 1: Colour by
#>            ^