USSC Confluence Version History and Log

USSC Confluence Version History and Log

ussc_confluence_version_history(id = id,
  username = Sys.getenv("CONFLUENCE_USERNAME"),
  password = Sys.getenv("CONFLUENCE_PASSWORD"))

Arguments

id
Page ID - a number found in the confluence URL

username
Your Confluence username which should be identical to your email. Defaults to an entry in .renviron file called CONFLUENCE_USERNAME.

password
Your Confluence API key (Get from https://confluence.atlassian.com/cloud/api-tokens-938839638.html). Defaults to an entry in .renviron file called CONFLUENCE_PASSWORD.

Examples

ussc_confluence_version_history(id = "942637057")

#> Please note: depending on how many edits have been made to the page, scraping the version history will take a while (~2 - 4 mins)... 
#>  

#> Joining, by = "version_history"

#> # A tibble: 315 x 16
#>    date       version_history message email name  title program status
#>    <date>               <dbl> <chr>   <chr> <chr> <chr> <chr>   <chr> 
#>  1 2019-04-01              41 ""      susa… Susa… (Par… I&E     Origi…
#>  2 2019-05-28              57 ""      susa… Susa… (Par… I&E     Origi…
#>  3 2019-05-29              60 ""      clai… Clai… (Par… I&E     Origi…
#>  4 2019-05-29              63 ""      susa… Susa… (Par… I&E     Origi…
#>  5 2019-06-16              71 ""      drew… Drew… (Par… I&E     Origi…
#>  6 2019-06-19              72 ""      drew… Drew… (Par… I&E     Origi…
#>  7 2019-07-10              79 ""      susa… Susa… (Par… I&E     Origi…
#>  8 2019-07-22              82 ""      drew… Drew… (Par… I&E     Origi…
#>  9 2019-08-17              88 ""      susa… Susa… (Par… I&E     Origi…
#> 10 2019-04-01              41 ""      susa… Susa… (Par… I&E     Draft…
#> # … with 305 more rows, and 8 more variables: original_publication_date <chr>,
#> #   new_publication_date <chr>, reasons_for_delay <chr>, author <chr>,
#> #   approver <chr>, type <chr>, production_coordinator <chr>,
#> #   publication_date <chr>